unit tests

+++ app/src/main/java/ferreira/hallefy/easyinvestment/domain/model/SimulationRequestBusiness.kt
@@ -1,7 +1,7 @@
 package ferreira.hallefy.easyinvestment.domain.model
 
 data class SimulationRequest(
-    var investedAmount: Int,
+    var investedAmount: Double,
     var index: String,
     var rate: Int,
     var isTaxFree: Boolean,
